Add mapping class

+++ app/src/main/java/com/example/myweatherappadvanced/calculate/Calculator.java
@@ -1,15 +1,15 @@
 package com.example.myweatherappadvanced.calculate;
 
 public class Calculator {
-    public static int cToF(int valC) {
+    public static long cToF(long valC) {
         return (int) Math.round(valC * 1.8 + 32);
     }
 
-    public static int gPaToMm(int valGPa) {
+    public static long gPaToMm(long valGPa) {
         return (int) Math.round(valGPa * 0.75);
     }
 
-    public static int msToKmh(int valMs) {
+    public static long msToKmh(float valMs) {
         return (int) Math.round(valMs * 3.6);
     }
 }
